Rename components' names to pascal case
